requirements.txt
dashscope==1.14.0
google-generativeai==0.4.1
numpy==1.26.3
openai==1.42.0
pandas==2.2.2
protobuf==4.25.3
python_Levenshtein==0.25.1
scikit_learn==1.4.1.post1
scipy==1.14.1
sentence_transformers==2.6.0
spacy==3.7.4
tqdm==4.66.1
zhipuai==2.0.1
